feat: seprate components into files
