+++ packages/frontend/src/view/components/status/LastDiscoveredBlock.tsx
@@ -1,0 +1,24 @@
+import { Row } from './Row'
+import { SubsectionHeader } from './SubsectionHeader'
+
+export function LastDiscoveredBlock({
+  blockNumber,
+}: {
+  blockNumber: number | null
+}) {
+  if (!blockNumber) {
+    return (
+      <SubsectionHeader
+        title="Last Discovered block"
+        subtitle="Data is missing"
+      />
+    )
+  }
+
+  return (
+    <>
+      <SubsectionHeader title="Last Discovered Block" />
+      <Row label="Last discovered block" value={blockNumber} />
+    </>
+  )
+}
